Latest031123

+++ Do IT/Do IT/RemoveLocation.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Data.SQLite;
+using System.Text.RegularExpressions;
 
 namespace Do_IT
 {
@@ -26,26 +27,33 @@ namespace Do_IT
 
         private void RemoveButton_Click(object sender, EventArgs e)
         {
-            Forms.conn.Open();
-            SQLiteCommand sql = new SQLiteCommand($"SELECT Isle, Bay FROM ValidLocations WHERE Isle = '{IsleTextBox.Text}' AND Bay = '{BayTextBox.Text}'", Forms.conn);
-            SQLiteDataReader reader = sql.ExecuteReader();
-
-            if (reader.Read())
+            if (Regex.IsMatch(IsleTextBox.Text, RegExFormats.anynumber) && Regex.IsMatch(BayTextBox.Text, RegExFormats.anynumber))
             {
-                SQLiteCommand sql2 = new SQLiteCommand($"DELETE FROM ValidLocations WHERE Isle = '{IsleTextBox.Text}' AND Bay = '{BayTextBox.Text}'", Forms.conn);
-                sql2.ExecuteNonQuery();
-                SQLiteCommand sql3 = new SQLiteCommand($"DELETE FROM ProductLocations WHERE Isle = '{IsleTextBox.Text}' AND Bay = '{BayTextBox.Text}'", Forms.conn);
-                sql3.ExecuteNonQuery();
-                SQLiteCommand sql4 = new SQLiteCommand($"UPDATE Products SET Located = 2 WHERE Located = 1 AND Barcode NOT IN (SELECT DISTINCT Barcode FROM ProductLocations)", Forms.conn);
-                sql4.ExecuteNonQuery();
-                MessageBox.Show("Location Removed");
+                Forms.conn.Open();
+                SQLiteCommand sql = new SQLiteCommand($"SELECT Isle, Bay FROM ValidLocations WHERE Isle = '{IsleTextBox.Text}' AND Bay = '{BayTextBox.Text}'", Forms.conn);
+                SQLiteDataReader reader = sql.ExecuteReader();
+
+                if (reader.Read())
+                {
+                    SQLiteCommand sql2 = new SQLiteCommand($"DELETE FROM ValidLocations WHERE Isle = '{IsleTextBox.Text}' AND Bay = '{BayTextBox.Text}'", Forms.conn);
+                    sql2.ExecuteNonQuery();
+                    SQLiteCommand sql3 = new SQLiteCommand($"DELETE FROM ProductLocations WHERE Isle = '{IsleTextBox.Text}' AND Bay = '{BayTextBox.Text}'", Forms.conn);
+                    sql3.ExecuteNonQuery();
+                    SQLiteCommand sql4 = new SQLiteCommand($"UPDATE Products SET Located = 2 WHERE Located = 1 AND Barcode NOT IN (SELECT DISTINCT Barcode FROM ProductLocations)", Forms.conn);
+                    sql4.ExecuteNonQuery();
+                    MessageBox.Show("Location Removed");
+                }
+                else
+                {
+                    MessageBox.Show("Invalid Location");
+                }
+                reader.Close();
+                Forms.conn.Close();
             }
             else
             {
-                MessageBox.Show("Invalid Location");
+                MessageBox.Show("Invalid input");
             }
-            reader.Close();
-            Forms.conn.Close();
         }
     }
 }
